+++ lib/src/api/entities/transaction/transaction_type.dart
@@ -1,1 +1,1 @@
-enum TransactionType { deposit, withdrawal, transfer }
+enum TransactionType { deposit, withdrawal, transferIn, transferOut }